+++ spec/Pim/Bundle/TransformBundle/Normalizer/Flat/MediaNormalizerSpec.php
@@ -66,7 +66,7 @@ class MediaNormalizerSpec extends ObjectBehavior
 
     function it_normalizes_media_by_using_file_and_export_path_to_prepare_the_copy(AbstractProductMedia $media, $manager)
     {
-        $media->getFilePath()->willReturn('/tmp/file/foo.jpg');
+        $manager->getFilePath($media)->willReturn('/tmp/file/foo.jpg');
         $manager->getExportPath($media)->willReturn('files/sku/attribute/foo.jpg');
 
         $this